refactor(ml): one AgentResponseParser for LLM JSON extraction
Candidate #4 from /improve-codebase-architecture. The fence-strip +
balanced-brace extraction (the part fixed twice for the 2026-05-16
trailing-prose regression) was duplicated character-for-character in
GuitarAlchemistAgentBase and AgentSkillBase. Concentrate it in one
internal AgentResponseParser.ExtractJsonCandidate; both bases keep
their own deserialize + fallback mapping. Behaviour-preserving.
